Add files via upload
Executive Sales Report
You are an analyst working at a major retailer. The sales director has approached you and asked for you to prepare a report that he can take to the executives regarding sales performance and to make sure it looks good. Unfortunately, he is unavailable for any further conversations about what he wants in the report!!


OBJECTIVE: Produce a sales report for executives.